requirements.txt
python-telegram-bot
python-decouple
requests
